Split services query-params.js
